+++ CHANGES.rst
@@ -8,6 +8,9 @@ Fixed:
 
 - Move blob-cache setting to prod cfg #4739
   [slindner]
+- Show comments in archive. #4819
+  [tlotze]
+
 
 
 1.9.2 (04.08.22)